fs: limit filesystem stacking depth
CVE-2016-1583

Add a simple read-only counter to super_block that indicates how deep this
is in the stack of filesystems.  Previously ecryptfs was the only stackable
filesystem and it explicitly disallowed multiple layers of itself.

Overlayfs, however, can be stacked recursively and also may be stacked
on top of ecryptfs or vice versa.

To limit the kernel stack usage we must limit the depth of the
filesystem stack.  Initially the limit is set to 2.

Change-Id: I78e6096cbcaf0515307630a6e8cb8d86963a0b79
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>