driver: sound: fix CVE-2016-6690
CVE-2016-6690

A denial of service vulnerability in the kernel could allow a local
malicious application to cause a device reboot.
There is no validation of the codec variable passed to the snd_soc_read
and snd_soc_write functions.
The fix is designed to add a check for null function pointers in the
dummy sound driver.

Change-Id: I970a46b17fe3d54af37699b2d3a1cb83b9227528
Signed-off-by: Lianchao Song <songlianchao@huawei.com>