driver: binder: fix CVE-2016-6689
CVE-2016-6689

An information disclosure vulnerability in Binder could enable a local
malicious application to access data outside of its permission levels.
The interaction between the kernel /dev/binder and the usermode
Parcel.cpp means that when a Binder object is passed as BINDER_TYPE_BINDER
or BINDER_TYPE_WEAK_BINDER, a pointer to that object (in the server process)
is leaked to the client process as the cookie value.
This leads to a leak of a heap address in many of the privileged Binder
services, including system_server.
The fix is designed to zero out the Binder pointer and cookie before sending
it to the client process.

Change-Id: Ic8f21578b4a5889813ce66af15002c9d7ae8f6cf
Signed-off-by: Lianchao Song <songlianchao@huawei.com>
(cherry picked from commit 1ddb21939ba4ff73e09335c63b59c02064fe5ce6)