Revert "ecryptfs: forbid opening files without mmap handler"
CVE-2016-1583

This reverts commit 2f36db71009304b3f0b95afacd8eba1f9f046b87.

It fixed a local root exploit but also introduced a dependency on
the lower file system implementing an mmap operation just to open a file,
which is a bit of a heavy hammer.  The right fix is to have mmap depend
on the existence of the mmap handler instead.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Cc: stable@vger.kernel.org
Signed-off-by: Tyler Hicks <tyhicks@canonical.com>

Change-Id: Ifa70e768b6bdd0725979fb7cc5c5471432307c8b
Signed-off-by: Lianchao Song <songlianchao@huawei.com>
(cherry picked from commit a7ba5db519b7bd99ca8acd54648db19fc7b20583)